(feat) Enable modifying of part in viewWorkoutModal
